+++ TableProTests/Editor/PasteHighlightCancelTests.swift
@@ -187,7 +187,9 @@ struct PasteHighlightCancelTests {
             numbers: EditorTheme.Attribute(color: Self.numbers),
             strings: EditorTheme.Attribute(color: .systemRed),
             characters: EditorTheme.Attribute(color: .systemRed),
-            comments: EditorTheme.Attribute(color: .systemGreen)
+            comments: EditorTheme.Attribute(color: .systemGreen),
+            operators: EditorTheme.Attribute(color: .systemBrown),
+            functions: EditorTheme.Attribute(color: .systemIndigo)
         )
         let controller = TextViewController(
             string: Self.document,